Update import paths for link actions
